+++ packages/init/.gitignore
@@ -1,2 +1,1 @@
-*.js
-types/*.js
+lib/
